refactored auth
