+++ src/app/login/page.tsx
@@ -1,74 +1,8 @@
-"use client";
+import { getSessionServer } from "@/lib/server/session";
+import { Session } from "@supabase/auth-helpers-nextjs";
+import LoginComp from "./components/LoginForm";
 
-import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
-import { useRouter } from "next/navigation";
-import { useState } from "react";
-
-import type { Database } from "@/types/types";
-
-export default function Login() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const router = useRouter();
-  const supabase = createClientComponentClient<Database>();
-
-  const handleSignUp = async () => {
-    const { data, error } = await supabase.auth.signUp({
-      email,
-      password,
-      options: {
-        emailRedirectTo: `${location.origin}/auth/callback`,
-      },
-    });
-
-    console.log(data);
-    router.refresh();
-  };
-
-  const handleSignIn = async () => {
-    const { data, error } = await supabase.auth.signInWithPassword({
-      email,
-      password,
-    });
-
-    console.log(data);
-    router.refresh();
-  };
-
-  const handleSpotify = async () => {
-    const { data, error } = await supabase.auth.signInWithOAuth({
-      provider: "spotify",
-      options: {
-        redirectTo: `${location.origin}/auth/callback`,
-      },
-    });
-
-    console.log(data);
-    router.refresh();
-  };
-
-  const handleSignOut = async () => {
-    await supabase.auth.signOut();
-    router.refresh();
-  };
-
-  return (
-    <>
-      <input
-        name="email"
-        onChange={(e) => setEmail(e.target.value)}
-        value={email}
-      />
-      <input
-        type="password"
-        name="password"
-        onChange={(e) => setPassword(e.target.value)}
-        value={password}
-      />
-      <button onClick={handleSignUp}>Sign up</button>
-      <button onClick={handleSignIn}>Sign in</button>
-      <button onClick={handleSpotify}>Spotify</button>
-      <button onClick={handleSignOut}>Sign out</button>
-    </>
-  );
+export default async function Login() {
+  const session = (await getSessionServer()) as Session;
+  return <LoginComp session={session} />;
 }
